+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - stateless http security configuration for `/auth` route
   - add spring AuthtenticationManager to handle authentications
   - implement `User` and `AuthorizationProfile` based on spring security models
+  - implement IUserReposity following JpaRepository interface
 
 ## 2021.11.17